Added Join Page and Contact Page
In App and Header, I made the Header sticky, this means it will always stay
on top of the screen even when you scroll past it. I added a div to push
the remaining content down. I also editted the style of the onFocus feature
for all of our buttons. They are now big and blue. This helps any user who
wants to use tabs to navigate through the page.

In the Join and Contact Page, I made everything into a form. The submit button
is now an input button in the form. This allows me to use the required tag
to add validation for me. When the user submits, the data does not go anywhere,
as there is no where to send any information. But it sends an alert to the user
and moves the user back to the landing page.